Added two player functionality
*VARIABLE CHANGES*
'currentPlayer' is used to keep track of which player is currently going.
'currentShape' replaces 'requestedShape' because the shape is now associated with the player and handled by the script.


*HANDLER CHANGES*
'NewGame' now makes the board clear, resets the values of 'currentShape', and 'currentPlayer'. It also tells the players that it is player 1's turn and they are starting on X. 

*FUNCTION CHANGES*
Removed helpers for 'changeBoard' because 'currentShape' makes them redundant. 
'updatePlayer' changes the value of 'currentPlayer', and 'currentShape' to be the opposite of what they currently are.