Complete 2D array and Fallback Functionality
*FUNCTION CHANGES*
All functions changed to handle the board as a 2d array instead of three 1d arrays.

*HANDLER CHANGES*
'AMAZON.FallbackIntent' has been added to give a message when a null value is the result of an utterance. This prevents the skill from exiting with an error when the user says something unplanned.